device_id

Function encode_device_numbers

pub fn encode_device_numbers(major: u32, minor: u32) -> u64

Encodes the major and minor numbers as a u64 value.

The lower 32 bits use the same encoding strategy as Linux. See the Linux implementation at: https://github.com/torvalds/linux/blob/0ff41df1cb268fc69e703a08a57ee14ae967d0ca/include/linux/kdev_t.h#L39-L44.

If the major or minor device number is too large, the additional bits will be recorded using the higher 32 bits. Note that as of 2025, the Linux kernel still has no support for 64-bit device IDs: https://github.com/torvalds/linux/blob/0ff41df1cb268fc69e703a08a57ee14ae967d0ca/include/linux/types.h#L18. So this encoding follows the implementation in glibc: https://github.com/bminor/glibc/blob/632d895f3e5d98162f77b9c3c1da4ec19968b671/bits/sysmacros.h#L26-L34.